Create  index.html
This HTML file includes a paragraph element (<p>) with an id of "output", which is where the output of the isChatGptText function will be displayed. The <script> element includes the JavaScript code, and the last line of the JavaScript code sets the innerHTML property of the paragraph element to the output of the function.

Signed-off-by: alexcavuilati <124851440+alexcavuilati@users.noreply.github.com>